requirements.txt
# FastAPI and server
fastapi==0.121.2
uvicorn==0.38.0
python-multipart==0.0.20

# Google Cloud / Firebase
google-cloud-firestore
google-cloud-pubsub
firebase-admin

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-dotenv==1.2.1

# Data validation
pydantic==2.12.4
pydantic_core==2.41.5
email-validator==2.3.0

# HTTP client
httpx==0.28.1

# Core dependencies
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.0
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
idna==3.11
sniffio==1.3.1
starlette==0.49.3
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
watchfiles==1.1.1

# Web Framework
fastapi==0.121.2
uvicorn[standard]==0.34.0

# Database
google-cloud-firestore==2.19.0

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
PyJWT==2.10.1

# Configuration
python-dotenv==1.0.1

# Data Validation
pydantic==2.10.3
email-validator==2.2.0

advertools==0.17.1
aiofiles==25.1.0
aiolimiter==1.2.1
aiosmtplib==5.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Automat==25.4.16
bcrypt==5.0.0
beautifulsoup4==4.14.2
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
edgartools==4.33.0
email-validator==2.3.0
fastapi==0.121.2
filelock==3.20.0
firebase_admin==7.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hishel==0.1.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpxthrottlecache==0.2.1
humanize==4.14.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.12.2
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
nest-asyncio==1.6.0
numpy==2.3.5
oauthlib==3.3.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parsel==1.10.0
passlib==1.7.4
Protego==0.5.0
proto-plus==1.26.1
protobuf==6.33.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyDispatcher==2.0.7
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
queuelib==1.8.0
rank-bm25==0.2.2
RapidFuzz==3.14.3
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
Scrapy==2.13.4
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stamina==25.1.0
starlette==0.49.3
tabulate==0.9.0
tenacity==9.1.2
textdistance==4.6.3
tldextract==5.3.0
tqdm==4.67.1
Twisted==25.5.0
twython==3.9.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.38.0
w3lib==2.3.1
watchfiles==1.1.1
websockets==15.0.1
zope.interface==8.1.1
lxml==5.1.0
openai==1.59.5
python-dotenv==1.0.0
